Fix types and comment secureChannel so CAP can be loaded on card

+++ applet/src/main/java/applet/Record.java
@@ -89,7 +89,7 @@ public class Record {
      * @throw StorageException
      * */
     private void setSecret(byte[] buffer, short secretOffset, byte secretLength) throws InvalidArgumentException, StorageException {
-        if (secretLength < SECRET_MIN_LEN || secretLength > SECRET_MAX_LEN || buffer.length < secretOffset + SECRET_MAX_LEN) {
+        if (secretLength < SECRET_MIN_LEN || secretLength > SECRET_MAX_LEN || buffer.length < (short)(secretOffset + SECRET_MAX_LEN)) {
             throw new InvalidArgumentException();
         }
         try {
@@ -112,7 +112,7 @@ public class Record {
      * @return length of concatenated names and their lengths
      * */
     public short getSecret(byte[] outputBuffer, short outputOffset) throws InvalidArgumentException, ConsistencyException {
-        if (outputBuffer.length < SECRET_MIN_LEN || outputBuffer.length < outputOffset + SECRET_MAX_LEN) {
+        if (outputBuffer.length < SECRET_MIN_LEN || outputBuffer.length < (short)(outputOffset + SECRET_MAX_LEN)) {
             throw new InvalidArgumentException();
         }
